- Image asset restructuring (Production/ + organized Org/)
- Overflow protection (SafeMath, int→long promotion)
- Hardcoding removal + dead code cleanup
- Sound theme system (4 presets, per-sound volume control)
- Batch 2 monsters, sprite tools, balance docs

+++ Windows/StatisticsWindow.xaml.cs
@@ -517,7 +517,7 @@ namespace DeskWarrior.Windows
             // Summary 계산
             int totalRecords = records.Count;
             double avgDamage = records.Average(r => r.FinalDamage);
-            int maxDamage = records.Max(r => r.FinalDamage);
+            long maxDamage = records.Max(r => r.FinalDamage);
 
             TxtDmRecordCount.Text = totalRecords.ToString();
             TxtDmAvgDamage.Text = FormatNumber((long)avgDamage);
